+++ requests_auth/__init__.py
@@ -1,2 +1,2 @@
-from .authentication import Basic, HeaderApiKey, QueryApiKey, NTLM, OAuth2, MicrosoftOAuth2 as MSOAuth2, Auths
+from .authentication import Basic, HeaderApiKey, QueryApiKey, NTLM, OAuth2, AzureActiveDirectory as AzureAD, Okta, Auths
 from .oauth2_tokens import JsonTokenFileCache